Duy_fix_VerifyCode of account and CreateOrderBuy have List product

+++ CameraServicesPlatform/CameraServicesPlatform.BackEnd.Common/DTO/Request/OrderDetailRequest.cs
@@ -6,11 +6,11 @@ namespace CameraServicesPlatform.BackEnd.Common.DTO.Request
     {
         public Guid OrderID { get; set; }
         public Guid ProductID { get; set; }
-        public decimal ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
 
         [MaxLength(255)]
         public string ProductQuality { get; set; }
-        public decimal Discount { get; set; }
-        public decimal ProductPriceTotal { get; set; }
+        public double Discount { get; set; }
+        public double ProductPriceTotal { get; set; }
     }
 }
